add database exception handling for delete and update operations

+++ src/main/java/com/educandofe/course/services/UserService.java
@@ -4,7 +4,9 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import com.educandofe.course.services.exception.DatabaseException;
 
 import com.educandofe.course.Model.UserModel;
 import com.educandofe.course.repositorys.UserRepository;
@@ -31,13 +33,24 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        try {
+            if (!userRepository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Cannot delete user because there are related records");
+        }
     }
 
     public UserModel update(Long id, UserModel user) {
-        UserModel entity = userRepository.getReferenceById(id);
-        updateData(entity, user);
-        return userRepository.save(entity);
+        try {
+            UserModel entity = userRepository.getReferenceById(id);
+            updateData(entity, user);
+            return userRepository.save(entity);
+        } catch (jakarta.persistence.EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
     private void updateData(UserModel entity, UserModel user) {
         entity.setName(user.getName());
